+++ src/main/java/domain/AttendanceDate.java
@@ -1,7 +1,5 @@
 package domain;
 
-import domain.utils.DateTimeUtil;
-
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
