+++ src/test/java/domain/utils/DateTimeUtilTest.java
@@ -1,5 +1,6 @@
 package domain.utils;
 
+import global.utils.DateTimeUtil;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDate;
